Move utils folder into components/playlist
